// axportal_frontend/src/components/agents/builder/types/Agents.ts
export const NodeType = {
  Note: {
    name: 'note',
    title: 'Note',
    enable_config: false,
  },
  Input: {
    name: 'input__basic',
    title: 'Input',
    enable_config: false,
  },
  OutputFormatter: {
    name: 'output__chat',
    // name: 'output__formatter',
    title: 'Output Chat',
    enable_config: false,
  },
  OutputSelector: {
    name: 'output__keys',
    // name: 'output__selector',
    title: 'Output Keys',
    enable_config: false,
  },
  AgentGenerator: {
    name: 'agent__generator',
    title: 'Generator',
    enable_config: true,
  },
  AgentReactor: {
    name: 'agent__reactor',
    title: 'ReACTor',
    enable_config: false,
  },
  AgentCategorizer: {
    name: 'agent__categorizer',
    title: 'Categorizer',
    enable_config: false,
  },
  AgentCondition: {
    name: 'condition',
    title: 'Condition',
    enable_config: false,
  },
  AgentCoder: {
    name: 'agent__coder',
    title: 'Code',
    enable_config: false,
  },
  AgentMerger: {
    name: 'merger',
    title: 'Merger',
    enable_config: false,
  },
  AgentApp: {
    name: 'agent__app',
    title: 'AgentApp',
    enable_config: false,
  },
  RewriterHyDE: {
    name: 'retriever__rewriter_hyde',
    title: 'Rewriter HyDE',
    enable_config: false,
  },
  RewriterMultiQuery: {
    name: 'retriever__rewriter_multiquery',
    title: 'Rewriter MultiQuery',
    enable_config: false,
  },
  RetrieverRetriever: {
    // name: 'retriever__main',
    name: 'retriever__knowledge',
    title: 'Retriever',
    enable_config: false,
  },
  RetrieverReRanker: {
    name: 'retriever__doc_reranker',
    title: 'Doc ReRanker',
    enable_config: false,
  },
  RetrieverCompressor: {
    name: 'retriever__doc_compressor',
    title: 'Doc Compressor',
    enable_config: false,
  },
  RetrieverFilter: {
    name: 'retriever__doc_filter',
    title: 'Doc Filter',
    enable_config: false,
  },
  AgentReviewer: {
    name: 'agent__reviewer',
    title: 'Reviewer',
    enable_config: true,
  },
  Tool: {
    name: 'tool',
    title: 'Tool',
    enable_config: false,
  },
  AgentUnion: {
    name: 'union',
    title: 'Union',
    enable_config: false,
  },
  NewNode: {
    name: 'new_node',
    title: 'New Node',
    enable_config: false,
  },
  DUMMY: {
    name: 'dummy',
    title: 'Dummy',
    enable_config: false,
  },
};
// ...
export interface CustomNodeInnerData extends Record<string, unknown> {
  title?: string;
  isRun: boolean;
  isDone?: boolean;
  isError?: boolean;
  isToggle: boolean;
  logData?: string[];

  [key: string]: unknown;
}
// ...
export interface key {
  name: string;
  value: string | any | null;
  node_type: string;
  node_id: string;
}
// ...
export interface query_rewriter {
  llm_chain: llm_chain;
  include_ori_query: boolean;
}
// ...
export type CustomNode = Node & {
  id: string;
  type: string;
  position: { x: number; y: number };
  source_position: string | null;
  target_position: string | null;
  data: { innerData: CustomNodeInnerData } & (
    | InputNodeDataSchema
    | OutputKeysDataSchema
    | OutputChatDataSchema
    | RetrieverDataSchema
    | QueryRewriterDataSchema
    | ContextRefinerDataSchema
    | GeneratorDataSchema
    | AgentAppDataSchema // 20250913 추가
    | CoderDataSchema
    | ReACTorDataSchema
    | CategorizerDataSchema
    | ConditionDataSchema
    | UnionDataSchema
    | ReviewerDataSchema // 20250913 추가
    | ToolDataSchema
    | ReRankerDataSchema
    | Record<string, unknown>
  );
};

export type InputKeyItem = {
  name: string;
  required: boolean;
  description?: string;
  keytable_id: string | null;
  fixed_value: string | null;
};

export type OutputKeyItem = {
  name: string;
  keytable_id: string;
};

export type InputNodeDataSchema = {
  type: string | null;
  id: string | null;
  name: string;
  description: string | null;
  input_keys: InputKeyItem[];
};

export type OutputKeysDataSchema = {
  type: string | null;
  id: string | null;
  name: string;
  description: string | null;
  input_keys: InputKeyItem[];
};

// 20250913 추가
export type UnionDataSchema = {
  type: string | null;
  id: string | null;
  name: string;
  description: string | null;
  format_string: string | null;
  input_keys: InputKeyItem[] | null;
  output_keys: OutputKeyItem[] | null;
};

export type OutputChatDataSchema = {
  type: string | null;
  id: string | null;
  name: string;
  description: string | null;
  format_string: string | null;
};

export type RetrieverDataSchema = {
  type: string | null;
  id: string | null;
  retriever_id: string | null;
  name: string | null;
  description: string | null;
  kind: string;
  knowledge_retriever: KnowledgeRetriever;
  input_keys: InputKeyItem[];
  output_keys: OutputKeyItem[];
};

export type ReviewerDataSchema = {
  type: string | null;
  id: string | null;
  name: string | null;
  description: string | null;
  serving_name: string;
  serving_model: string;
  prompt_id: string;
  max_review_attempts: number;
  input_keys: InputKeyItem[];
  output_keys: OutputKeyItem[];
};

export type QueryRewriterDataSchema = {
  type: string | null;
  id: string | null;
  retriever_id: string | null;
  name: string | null;
  description: string | null;
  kind: string;
  query_rewriter: object;
  input_keys: InputKeyItem[];
  output_keys: OutputKeyItem[];
};
// ...
export type ContextRefinerDataSchema = {
  type: string | null;
  id: string | null;
  name: string | null;
  retriever_id: string | null;
  description: string | null;
  kind: string;
  context_refiner: object | null;
  input_keys: InputKeyItem[];
  output_keys: OutputKeyItem[];
};

export type GeneratorDataSchema = {
  type: string | null;
  id: string | null;
  name: string | null;
  description: string | null;
  serving_name: string;
  serving_model: string;
  prompt_id: string;
  fewshot_id: string | null;
  tool_ids: Array<string> | null;
  tools: Tool[];
  mcp_selections?: any[] | null; // MCP 선택 정보 추가
  input_keys: InputKeyItem[];
  output_keys: OutputKeyItem[];
};

// 20250913 추가
export type AgentAppDataSchema = {
  type: string | null;
  id: string | null;
  name: string | null;
  description: string | null;
  agent_app_id: string;
  api_key: string;
  input_keys: InputKeyItem[];
  output_keys: OutputKeyItem[];
};
// ...
export type CoderDataSchema = {
  type: string | null;
  id: string | null;
  name: string | null;
  description: string | null;
  code: string;
  input_keys: InputKeyItem[] | null;
  output_keys: OutputKeyItem[] | null;
};

export type ReACTorDataSchema = {
  type: string | null;
  id: string | null;
  name: string | null;
  description: string | null;
  serving_name: string;
  serving_model: string;
  prompt_id: string;
  fewshot_id: string | null;
  tool_ids: string[] | null;
  tools: Tool[];
  input_keys: InputKeyItem[];
  output_keys: OutputKeyItem[];
};

export type CategorizerDataSchema = {
  type: string | null;
  id: string | null;
  name: string | null;
  description: string | null;
  serving_name: string;
  serving_model: string;
  prompt_id: string;
  categories: Array<{
    category: string;
    description: string;
  }> | null;
  input_keys: InputKeyItem[];
  output_keys: OutputKeyItem[];
};

export type ConditionItem = {
  id: string;
  type: string;
  operator: string;
  input_key: {
    name: string;
    required: boolean;
    keytable_id: string | null;
  };
  value: {
    name: string;
    required: boolean;
    fixed_value?: string | null;
    keytable_id: string | null;
  };
};

export type ConditionDataSchema = {
  type: string | null;
  id: string | null;
  name: string;
  description: string | null;
  conditions: Array<ConditionItem>;
  default_condition: string;
  input_keys: InputKeyItem[] | null;
  output_keys: OutputKeyItem[] | null;
};

export type MergerDataSchema = {
  type: string | null;
  id: string | null;
  name: string;
  description: string | null;
  format_string: string | null;
  input_keys: InputKeyItem[] | null;
  output_keys: OutputKeyItem[] | null;
};

export type ToolDataSchema = {
  type: string | null;
  id: string | null;
  description: string | null;
  name: string | null;
  tool_id: string;
  input_keys: InputKeyItem[] | null;
  output_keys: OutputKeyItem[] | null;
};

export type ReRankerDataSchema = {
  type: string | null;
  id: string | null;
  description: string | null;
  name: string | null;
  retriever_id: string | null;
  kind: string;
  context_refiner: {
    rerank_cnf: {
      model_info: {
        serving_name: string | null;
        serving_model: string | null;
      };
    };
  };
  input_keys: InputKeyItem[] | null;
  output_keys: OutputKeyItem[] | null;
};

export type KnowledgeRetriever = {
  repo_id: string;
  repo_kind?: string;
  project_id?: string | null;
  active_collection_id?: string | null;
  index_name?: string | null;
  script?: string | null;
  serving_name: string | null;
  connect_type: string | null;
  connect_info: RetrieverApiInfo | RetrieverVectordbInfo | null;
  embedding_info?: Record<string, any> | null;
  knowledge_info?: Record<string, any> | null;
  vectordb_conn_info?: Record<string, any> | null;
  retrieval_options: RetrievalOptions | null;
};

export interface Category {
  id: string;
  category: string;
  description: string;
}

export interface Condition {
  id: string;
  type: string;
  operator: string;
  input_key: {
    name: string;
    required: boolean;
    keytable_id: string | null;
  };
  value: {
    name: string;
    required: boolean;
    fixed_value?: string | null;
    keytable_id: string | null;
  };
}